requirements.txt
beautifulsoup4
selenium
webdriver-manager
python-dotenv
jsonlines
